pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "manual-excel-loader"
version = "0.1.0"
description = "Excel/CSV/SQL → SQL/CSV loader with data validation for GreenPlum and ClickHouse"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1,<4.0",
    "python-dateutil>=2.9,<3.0",
    "tqdm>=4.66,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-cov>=4.1,<6.0",
    "ruff>=0.4,<1.0",
]

# Пакет теперь лежит в dags/, а не в src/
# Это позволяет копировать dags/ напрямую в $AIRFLOW_HOME/dags/
# без pip install — Airflow добавляет dags/ в sys.path автоматически.
[tool.setuptools.packages.find]
where = ["dags"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# dags/ добавлен в pythonpath чтобы тесты импортировали пакет без установки
pythonpath = ["dags"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.coverage.run]
source = ["dags/manual_excel_loader"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
